fixing type to int

+++ supervised_learning/object_detection/5-yolo.py
@@ -191,8 +191,8 @@ class Yolo():
         ni = len(images)
         input_h = self.model.input.shape[2]
         input_w = self.model.input.shape[1]
-        pimages=np.ndarray([ni,input_h,input_w,3])
-        image_shapes = np.ndarray([ni,2])
+        pimages=np.ndarray([ni,input_h,input_w,3],dtype=int)
+        image_shapes = np.ndarray([ni,2], dtype= int)
 
         for i in range(len(images)):
             resized_img = cv2.resize(images[i], (input_w,input_h), interpolation=cv2.INTER_CUBIC)
